BASE64: int -> size_t

+++ base64.c
@@ -4,17 +4,17 @@
 
 const char charset[64] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 
-int base64(const char* data, size_t len_data, char* res, size_t res_len) {
-    int unpadded_len = (len_data*8 + 5) / 6; // ceil(data_nb_bits/6)
-    int padding_len = (4 - unpadded_len) % 4;
-    int size_needed = unpadded_len + padding_len;
+size_t base64(const char* data, size_t len_data, char* res, size_t res_len) {
+    size_t unpadded_len = (len_data*8 + 5) / 6; // ceil(data_nb_bits/6)
+    size_t padding_len = (4 - unpadded_len) % 4;
+    size_t size_needed = unpadded_len + padding_len;
     if (res_len < size_needed) {
         printf("Insufficient size\n");
         return size_needed;
     }
 
-    int j = 0;
-    int i = 0;
+    size_t j = 0;
+    size_t i = 0;
     while (j + 2 < len_data) {
         res[i++] = charset[data[j + 0] >> 2];
         res[i++] = charset[((data[j + 0] & 0b00000011) << 4) | (data[j + 1] >> 4)];
@@ -45,13 +45,13 @@ int base64(const char* data, size_t len_data, char* res, size_t res_len) {
     return i;
 }
 
-int base64_str(const char* s, char* res, size_t res_len) {
+size_t base64_str(const char* s, char* res, size_t res_len) {
     base64(s, strlen(s), res, res_len);
 }
 
 int main(void) {
     char res[37];
-    int n = base64_str("Many hands make light work.", res, sizeof res);
+    size_t n = base64_str("Many hands make light work.", res, sizeof res);
     if (n >= sizeof res) {
         printf("%d: Size of res buffer is not enough: %d given and needs %d!\n", __LINE__, sizeof res, n);
         return 1;
